Updated service to search by sector instead of name

+++ apesc/src/main/java/com/example/apesc/service/employee/FuncionarioServiceImpl.java
@@ -1,13 +1,16 @@
 package com.example.apesc.service.employee;
 
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
+import com.example.apesc.exception.CustomException;
+import com.example.apesc.exception.ErrorConstants;
 import com.example.apesc.model.Funcionario;
 import com.example.apesc.repository.FuncionarioRepository;
 import com.example.apesc.util.FuncionarioValidation;
-
 import lombok.AllArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
 
 @Service
 @AllArgsConstructor
@@ -29,11 +32,17 @@ public class FuncionarioServiceImpl implements FuncionarioService {
     }
 
     public void delete(Long id) {
+        if (!funcionarioRepository.existsById(id)) {
+            throw new CustomException(
+                ErrorConstants.ID_NOT_FOUND,
+                HttpStatus.NOT_FOUND
+            );
+        }
         funcionarioRepository.deleteById(id);
     }
 
-    public Funcionario findByNome(String nome) {
-        return funcionarioRepository.findByNome(nome);
+    public List<Funcionario> findBySetor(String setor) {
+        return funcionarioRepository.findBySetorContainingIgnoreCase(setor);
     }
     
 }
